LeAudio: Fix crash on bad remote audio allocation config
If multiple devices in a group happen to have the same audio
allocation, we will not know which one to choose on the 2nd
device. Since this is abnormal we should print an error but
not crash. Return any left or any right allocation that the
device has. In worst case, the device has no left nor right
location and we will return Disallowed value of 0. In such
case, the remote device can deny the stream enablement.

Bug: 256967923
Test: atest --host bluetooth_le_audio_test bluetooth_le_audio_client_test bluetooth_test_broadcaster bluetooth_test_broadcaster_state_machine --no-bazel-mode
Change-Id: Ib3a3f68a29690e8c50683476a3acbf11e7e33938